🧰 refactored code
