+++ app/src/main/java/dev/spikeysanju/expensetracker/services/exportcsv/TransactionsCsv.kt
@@ -1,0 +1,33 @@
+package dev.spikeysanju.expensetracker.services.exportcsv
+
+import com.opencsv.bean.CsvBindByName
+import dev.spikeysanju.expensetracker.model.Transaction
+
+data class TransactionsCSV(
+    @CsvBindByName(column = "title")
+    val title: String,
+    @CsvBindByName(column = "amount")
+    val amount: Double,
+    @CsvBindByName(column = "transactionType")
+    val transactionType: String,
+    @CsvBindByName(column = "tag")
+    val tag: String,
+    @CsvBindByName(column = "date")
+    val date: String,
+    @CsvBindByName(column = "note")
+    val note: String,
+    @CsvBindByName(column = "createdAt")
+    val createdAtDate: String
+)
+
+fun List<Transaction>.toCsv() = map {
+    TransactionsCSV(
+        title = it.title,
+        amount = it.amount,
+        transactionType = it.transactionType,
+        tag = it.tag,
+        date = it.date,
+        note = it.note,
+        createdAtDate = it.createdAtDateFormat,
+    )
+}
